+++ NEWS.md
@@ -1,6 +1,7 @@
 # *baRulho 1.0.1*
 
 * New function align_test_files() to set time of signals in aligned re-recorded files
+* Parallel available on internal 'prep_X_bRlo_int()' function
 
 # *baRulho 1.0.0*